move from DSpace Bitstream class to DryadBitstream
